python2 again

+++ scripts/diameter.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
+if [ -z "$1" ] 
+then
+    echo "Please pass <n> as argument."
+    exit
+fi
+
 mkdir -p data
 
-rm data/diameter.txt 2> /dev/null
-sage ../sage/diameter.sage > data/diameter.txt
+sage ../sage/diameter.sage $1 >> data/diameter.txt
 rm ../sage/diameter.sage.py
